[WIP] Migrate existing CLI tests to new picocli implementation
